requirements.txt
django>=5.0,<6.0
tensorflow>=2.15.0
Pillow>=10.0
numpy>=1.24
